requirements.txt
numpy>=1.24.0
scipy>=1.10.0
openai==1.3.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
httpx==0.25.1
matplotlib>=3.7.0
redis>=4.6.0
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
alembic==1.12.1
pytest==7.4.3
aioredis==2.0.1
apscheduler>=3.10.1
pandas>=2.0.0
requests>=2.31.0
gitpython>=3.1.30
python-multipart>=0.0.6 